requirements.txt
requests
geopandas
google-genai
